requirements.txt
# Core image ops
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=9.0.0
tqdm>=4.60.0

# Phase 1: object localization (YOLOv11)
ultralytics>=8.3.0

# Stage 3: inference (torchvision + timm classifiers)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Stage 4: LPIPS evaluation
lpips>=0.1.4
